Ann new frontend styleguide
Okay, so we're just cleaning up the various markdown documents in here. You can see in @docs/archive/ There are a lot of old AI generated documents. Much of the content of these is not
 based in the current reality of this project. However, we are trying to create a couple of new documents in @docs/developer/ which should contain all of the evergreen documents. I want
to make sure that we have one in here about front-end styling. So I'd like you to look at the front-end app and the React components and make sure that you've written a good
comprehensive document which explains how the kind of design system visually works.

So things like colors, design style and so on. You're going to need to look at the code. You may want to consult the architecture guide but you may also want to have a look at the code.
I think there are a couple of documents in the archive that relate to this, but I'm not sure I trust them to be accurate. So please make sure this reflects the current reality. There's
no need to make these documents ridiculously long, just the information that AI coding agents are going to need when working with front-end code, CSS, Tailwind, etc. Once you've worked
this out, I want you to create a new document in... `docs/developer` which details this front end styling and design stuff.